requirements.txt
fastmcp>=2.0.0
httpx>=0.27.0
pyjwt>=2.8.0
yt-dlp>=2024.0.0
faster-whisper>=1.0.0
